remove the old uninstall hook
