+++ package.json
@@ -101,7 +101,6 @@
     "prettier:check": "prettier --check src/",
     "lint": "eslint . --max-warnings 0",
     "lint:fix": "eslint . --fix",
-    "vscode:uninstall": "node ./out/uninstall",
     "test": "npm run clear-out && tsc && npm run test:copyassets && node ./out/test/runTest.js",
     "vsce:package": "vsce package",
     "vsce:publish": "vsce publish",
@@ -159,7 +158,6 @@
   },
   "dependencies": {
     "await-semaphore": "^0.1.3",
-    "date-fns": "^2.19.0",
     "extract-zip": "^2.0.1",
     "https-proxy-agent": "^5.0.0",
     "semver": "^7.3.2",
