+++ webpack.config.js
@@ -11,8 +11,7 @@ const config = {
   target: "node", // vscode extensions run in a Node.js-context 📖 -> https://webpack.js.org/configuration/node/
 
   entry: {
-    extension: "./src/extension.ts",
-    uninstall: "./src/uninstall.ts"
+    extension: "./src/extension.ts"
   }, // the entry point of this extension, 📖 -> https://webpack.js.org/configuration/entry-context/
   output: {
     // the bundle is stored in the 'dist' folder (check package.json), 📖 -> https://webpack.js.org/configuration/output/
